Refactor the CLI so it can be used by other libraries
The CLI is a useful tool and if you are developing external
nodes you want to use it. This commit makes it so that is easier.

It takes the following approach

1. Split the CLI into a libcli which exports a function previously
   known as main with an added callback, and a new executable which
   just calls that
2. Have the new function have a registerCallback which you can pass
3. Add cmake targets for the library only
4. Add a couple of inlines which avoid duplicated symbols from headers
   so you can get this all to work

As a result, a library can write a 'test' cli which uses the entire
current CLI with one function injected with a couple of lines of cmake,
and the internal CLI still works just with one extra initialization
time call to an empty function.