Scaffolds better error handling
The thiserror crate will allow us to better handle errors as they are
emitted from various parts of the program. For now, I propose that we
just default to propagating all errors up to the level of
`ElementaryCliError`; but as the runtime gets more nuanced, we can
decide how to better handle non-critical errors.